moved lighting structs
